+++ graphics/internal/vao.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 class VAO {
 public:
     VAO();
@@ -8,7 +10,7 @@ public:
 public:
     unsigned int getHandle();
     void bind();
-    void attribPointer3f(unsigned, unsigned, unsigned);
+    void attribPointer3f(unsigned, unsigned, unsigned long);
     static void unbind();
 
 private:
